EPP architectural refactor (#781)
